Fixed octant transformations.

+++ src/uk/ac/cam/rds46/cgip/sv2/Point2D.java
@@ -1,10 +1,10 @@
 package uk.ac.cam.rds46.cgip.sv2;
 
 public class Point2D {
-	public final int x;
-	public final int y;
+	public final double x;
+	public final double y;
 	
-	public Point2D(int px, int py) {
+	public Point2D(double px, double py) {
 		x = px;
 		y = py;
 	}
@@ -22,11 +22,13 @@ public class Point2D {
 		return new Point2D(x - Q.x, y - Q.y);
 	}
 	
-	public int dot(Point2D Q) {
+	public double dot(Point2D Q) {
 		return (x * Q.x) + (y * Q.y);
 	}
 	
 	public double magnitude() {
-		return Math.sqrt((double)((x * x) + (y * y)));
+		return Math.sqrt((x * x) + (y * y));
 	}
+	
+	
 }
